--- src/services/validations/validateOpinion.js
@@ -1,5 +1,0 @@
-import { isEmpty } from 'lodash';
-
-const validateOpinion = ({ opinion }) => !isEmpty(opinion);
-
-export default validateOpinion;
