--- src/services/validations/validatePasswordEquality.js
@@ -1,3 +1,0 @@
-const validatePasswordEquality = ({ password, passwordConfirmation }) => password === passwordConfirmation;
-
-export default validatePasswordEquality;
